requirements.txt
Babel==2.5.1
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Babel==0.11.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pbr==3.1.1
pytz==2017.3
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
Werkzeug==0.12.2
